base game now works
displays player choice, computer choice, counts rounds (round stays the same
if the result is a tie) The score shows each round, shows the round result,
and displays if you won or lost the game in round result once it reaches
a round of 5. Once round is 5 it resets the game.